fixed integer instead of float price bug in products table

+++ migrations/sqls/20260424142927-products-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS products (
     id SERIAL PRIMARY KEY,
     name VARCHAR(30) NOT NULL,
-    price INTEGER NOT NULL DEFAULT 0,
+    price NUMERIC(10, 2) NOT NULL DEFAULT 0,
     category VARCHAR(50)
 );
